+++ NEWS.md
@@ -1,6 +1,7 @@
 # povcalnetR (development version)
 * Add ability to handle testing servers
 * Fix handling of JSON responses
+* Add feature to query poverty quantiles instead of poverty lines
 
 # povcalnetR 0.1.1
 Patch: Address RMD check issues related to API calls in vignettes